requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.37.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
opencv-python==4.6.0.66
orjson==3.8.2
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
psycopg2-binary==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
starlette==0.21.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2